+++ module/Batch/src/Command/DailyTaskStats/DailyTaskStatsFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Batch\Command\DailyTaskStats;
+
+class DailyTaskStatsFactory {
+    public function __invoke($controllers): DailyTaskStats
+    {
+        return new DailyTaskStats($controllers->get('faucetdev'));
+    }
+}