+++ module/Batch/src/Command/MinerPayments/MinerPayments.php
@@ -20,6 +20,14 @@ class MinerPayments extends Command
     protected TableGateway $mPaymentTbl;
 
     /**
+     * Withdraw Buff Table
+     *
+     * @var TableGateway $mBuffTbl
+     * @since 1.0.0
+     */
+    protected TableGateway $mBuffTbl;
+
+    /**
      * Security Tools
      *
      * @var SecurityTools $mSecTools
@@ -43,6 +51,7 @@ class MinerPayments extends Command
     public function __construct($mapper)
     {
         $this->mPaymentTbl = new TableGateway('faucet_miner_payment', $mapper);
+        $this->mBuffTbl = new TableGateway('faucet_withdraw_buff', $mapper);
 
         $this->mSecTools = new SecurityTools($mapper);
         $this->mTxTools = new TransactionTools($mapper);
@@ -64,6 +73,54 @@ class MinerPayments extends Command
             '=====================',
             'Sending Miner Payments @ ' . date('Y-m-d H:i:s', time()),
             '---------------------',
+        ]);
+
+        $paymentQueue = $this->mPaymentTbl->select(['state' => 'open']);
+        $totalPayments = $paymentQueue->count();
+        if($totalPayments == 0) {
+            $output->writeln([
+                '## ERROR - NO PAYMENTS IN QUEUE',
+            ]);
+            return Command::SUCCESS;
+        }
+
+        $totalPaid = 0;
+        foreach ($paymentQueue as $payment) {
+            // send coins to user
+            $newBalance = $this->mTxTools->executeTransaction($payment->amount_coin, false, $payment->user_idfs, $payment->id, $payment->coin.'-nanov2', $payment->shares_percent.'% of all shares on pool.', 0, false);
+            if($newBalance) {
+                $totalPaid+=$payment->amount_coin;
+                // Add Buff for daily withdrawal limit
+                $this->mBuffTbl->insert([
+                    'ref_idfs' => 0,
+                    'ref_type' => 'mining',
+                    'label' => $payment->coin.' Mining',
+                    'days_left' => 1,
+                    'days_total' => 1,
+                    'amount' => $payment->amount_coin,
+                    'created_date' => date('Y-m-d H:i:s', time()),
+                    'user_idfs' => $payment->user_idfs
+                ]);
+            } else {
+                $output->writeln([
+                    '## ERROR - PAYMENT '.$payment->id.' COULD NOT BE EXECUTED',
+                ]);
+            }
+        }
+
+        $output->writeln([
+            '-- Sent '.$totalPayments.' Payments',
+        ]);
+        $output->writeln([
+            '-- Paid '.$totalPaid.' Coins to Users',
+        ]);
+
+        // Flag payments as done
+        $this->mPaymentTbl->update(['state' => 'paid'],['state' => 'open']);
+
+        $output->writeln([
+            '-- Payment run completed successfully',
+            '=====================',
         ]);
 
         return Command::SUCCESS;