+++ module/Batch/src/Command/MinerPayments/MinerPaymentsFactory.php
@@ -3,7 +3,6 @@
 namespace Batch\Command\MinerPayments;
 
 class MinerPaymentsFactory {
-
     public function __invoke($controllers): MinerPayments
     {
         return new MinerPayments($controllers->get('faucetdev'));