+++ module/Batch/src/Command/MinerStats/MinerStatsFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Batch\Command\MinerStats;
+
+class MinerStatsFactory {
+    public function __invoke($controllers): MinerStats
+    {
+        return new MinerStats($controllers->get('faucetdev'));
+    }
+}