+++ module/Batch/src/Command/OfferwallStats/OfferwallStatsFactory.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Batch\Command\OfferwallStats;
+
+class OfferwallStatsFactory {
+    public function __invoke($controllers): OfferwallStats
+    {
+        return new OfferwallStats($controllers->get('faucetdev'));
+    }
+}